+++ install.sh
@@ -1096,6 +1096,7 @@ BASE_PACKAGES=(
     espeak
     libespeak-ng1
     icecast2
+    fail2ban
     ca-certificates
     libusb-1.0-0
     libusb-1.0-0-dev